refactor(Terminal):移动 taskStatus 定义的文件位置
