+++ web/src/stores/constant/terminalTaskStatus.ts
