+++ web/src/stores/terminal.ts
@@ -5,7 +5,7 @@ import type { Terminal } from '/@/stores/interface/index'
 import { buildTerminalUrl } from '/@/api/common'
 import { uuid } from '/@/utils/random'
 import { timeFormat } from '/@/utils/common'
-import { taskStatus } from '/@/components/terminal/constant'
+import { taskStatus } from '/@/stores/constant/terminalTaskStatus'
 import { ElNotification } from 'element-plus'
 import { i18n } from '/@/lang/index'
 import { closeHotUpdate, openHotUpdate } from '/@/utils/vite'
