+++ web/src/views/backend/module/components/commonDone.vue
@@ -100,7 +100,7 @@ import { moduleInstallState } from '../types'
 import { onRefreshTableData } from '../index'
 import { useTerminal } from '/@/stores/terminal'
 import FormItem from '/@/components/formItem/index.vue'
-import { taskStatus } from '/@/components/terminal/constant'
+import { taskStatus } from '/@/stores/constant/terminalTaskStatus'
 import { ElMessageBox } from 'element-plus'
 import { useI18n } from 'vue-i18n'
 import { dependentInstallComplete } from '/@/api/backend/module'
